requirements.txt
fastapi
uvicorn[standard]
httpx
openai
tiktoken
upstash-redis
numpy
pandas
python-dotenv
